Fixed silly Windows build problems.

+++ packer/packer.h
@@ -19,12 +19,12 @@
 #include "packer_extensions.h"
 
 #define SWAP8(x) x
-#define SWAP16(x) ((x & 0x00FF << 8) | (x & 0xFF00 >> 8))
+#define SWAP16(x) (short) ((x & 0x00FF << 8) | (x & 0xFF00 >> 8))
 #define SWAP32(x) ((x & 0x000000FF << 24) | \
 		               (x & 0x0000FF00 << 8)  | \
 									 (x & 0x00FF0000 >> 8)  | \
 									 (x & 0xFF000000 >> 24))
-#define SWAPFLOAT(x) ( (*((int *) &x) & 0x000000FF << 24) | \
+#define SWAPFLOAT(x) (float) ( (*((int *) &x) & 0x000000FF << 24) | \
                        (*((int *) &x) & 0x0000FF00 << 8) | \
                        (*((int *) &x) & 0x00FF0000 >> 8) | \
                        (*((int *) &x) & 0xFF000000 >> 24))
